add errors

+++ src/services/messages.service.ts
@@ -1,9 +1,8 @@
 import { ApiError } from "../errors/api.error";
 import { IMessage, IMessageParams, IUpdateMessage } from "../types/messageType";
 import {firebase, storage} from "../firebase";
-import {collection, deleteDoc, doc, getDoc, getDocs, setDoc, updateDoc } from "firebase/firestore";
+import {collection, deleteDoc, doc, getDoc, getDocs, setDoc, updateDoc, DocumentSnapshot, QuerySnapshot, DocumentData} from "firebase/firestore";
 import { getDownloadURL, ref, uploadBytes } from "firebase/storage";
-
 
 class MessagesService {
     public async send (dto: IMessage, files: Express.Multer.File[] ): Promise<void> {
@@ -36,7 +35,9 @@ class MessagesService {
 
         try {
             const messageRef = doc(firebase, `chats/${chatId}/messages/${messageId}`);
-            await updateDoc(messageRef, { ...dto, update: new Date(),});
+             const updateMessage = getDoc(messageRef);
+                 await updateDoc(messageRef, { ...dto, update: new Date(),});
+
         } catch (e) {
             const status = e.status || 500;
             throw new ApiError(e.message, status);
@@ -45,12 +46,17 @@ class MessagesService {
     public async getMessagesByChatId ( params: Partial <IMessageParams>): Promise<IMessage[]> {
         const {chatId} = params
         try {
-            const messagesRef = collection(firebase, 'chats', chatId, 'messages'); // Путь к подколлекции
-            const snapshot = await getDocs(messagesRef);
+            const messagesRef = collection(firebase, 'chats', chatId, 'messages');
+            const snapshot: QuerySnapshot<DocumentData> = await getDocs(messagesRef);
             const messages = snapshot.docs.map(doc =>({
                 messageId: doc.id,
                 ...doc.data() as IMessage
             }));
+            if (messages.length == 0)
+            {
+                throw new ApiError( 'Messages not found', 404);
+            }
+
             return messages
         } catch (e) {
             const status = e.status || 500;
@@ -63,6 +69,11 @@ class MessagesService {
 
         try {
             const messageRef = doc(firebase, `chats/${chatId}/messages/${messageId}`);
+            const message: DocumentSnapshot =  await getDoc (messageRef)
+            if (!message.exists())
+            {
+                throw new ApiError( 'message not found', 404);
+            }
             await deleteDoc(messageRef);
         } catch (e) {
             const status = e.status || 500;
